worker/requirements.txt
elasticsearch==8.19.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
pdfplumber==0.11.0
python-docx==1.1.2
python-pptx==0.6.23